requirements.txt
capstone==3.0.5
pyelftools==0.25
yapf==0.24.0
